I need to change the dominance structure
I need to recognize the existence of modules, and then
build dominance info per function. Currently, it's taking the
*module* to be an entry block, and then assuming there are no paths
from anyting to anything. Thus, A dominates B trivially holds:

> every path from entry to B passes through A

is indeed true! There are no paths, so it is true that every path from
entry to B passes through A...