[MLIR] [LangRef] Clarify `dialect-item` parsing
Draw attention to the fact that the grammar is written in such a way
that it ensures balanced parentheses. Also add a note on how external
consumers of MLIR could easily parse such a grammar.

We take the time to restructure the examples, and add clarification
about illegal attribute declarations.